requirements.txt
streamlit
pandas
requests
plotly
piexif
pillow